Migrate BindsCycleGraph
