+++ compiler-tests/src/test/data/dump/ir/cycles/BindsCycleGraph.kt
@@ -1,0 +1,15 @@
+// DONT_SORT_DECLARATIONS
+interface Foo
+
+@Inject class Bar(val fooProvider: Provider<Foo>) : Foo
+
+/**
+ * A component with a cycle in which a `@Binds` binding depends on the binding that has to be
+ * deferred.
+ */
+@DependencyGraph
+interface BindsCycleGraph {
+  fun bar(): Bar
+
+  @Binds fun foo(bar: Bar): Foo
+}
